完成东方电子图形解析
全图显示没有问题，开关变位其中有3中状态的模板，有待确认。k35这个css名称解析有问题，同样的内容，换一个名称，解析就OK。

+++ textsvg.h
@@ -13,8 +13,8 @@
 class TextSvg
 {
 public:
-	int			m_x;
-	int			m_y;
+	qreal			m_x;
+	qreal			m_y;
 	int			m_fontSize;
 	int			m_fontWidth;
 	int			m_fontHeight;
